added recomendation component
